Bind joint indices as floats so the wheels actually render
The skinned vertex layout declared joint indices as AttribType::Uint8
with uvec4 a_indices, on the argument that bgfx's GL backend selects
glVertexAttribIPointer from !isFloat(type) && !normalized so the shader
side must be an integer type to match. That reasoning is correct and the
result still did not work: the indices never arrived, every wheel vertex
read a garbage joint index, sampled outside the uniform array, and the
triangles were discarded. All four wheels were absent, not misplaced.

Storing them as floats puts the attribute on the same
glVertexAttribPointer path position, normal and uv already use and are
known to work on. int(a_indices.x) is unchanged. Cost is 12 bytes per
vertex on one shared 6856-vertex mesh, about 82 KB.

Isolated by bisection, after tagging the wheel swatches a colour used
nowhere else so wheel pixels could be counted. Lifting bones 1-4 left
the render byte-identical; lifting bone 0 raised the whole car; a
constant u_boneMatrices[1] raised the whole car including wheels; and
dropping the weight blend changed nothing. Everything except a_indices
was eliminated by measurement.

Showcase wheel pixels by azimuth go 92/68/400/3276 to 4096/4724/992/5184.
The old two-digit figures were never wheels: they are the mirror
housings, which share the tagged column.

This also fixes something unreported: the wheels have never rotated and
never took suspension travel. They were not skinned at all, so
computeWheelTransforms()'s spin angle and suspension offset have been
computed every frame and discarded for the life of the port.

T9 attributed this symptom to the camera and a banked track; T20 ruled
out geometry, normals, occlusion, lighting, culling and every skinning
input but stopped one step short. It survived because every individual
component was correct -- mesh, normals, palette, buffer contents, and a
varying type matching the binding bgfx documents. Only the composition
failed, and nothing looks at the composition except the picture.

ctest is bgfx-free so it cannot assert "the wheels drew"; the manual
check is recorded in PORT_PROGRESS.md and takes one sweep.

check_car_rig.py PASS, car_proportions.py 17/17, ctest 36/36.

+++ cpp-port/src/render/vertex_skinned.h
@@ -20,7 +20,7 @@ struct PosNormalUVSkinnedVertex {
     float x, y, z;
     float nx, ny, nz;
     float u, v;
-    uint8_t joints[4];
+    float joints[4];
     float weights[4];
 
     static bgfx::VertexLayout layout() {
@@ -29,7 +29,7 @@ struct PosNormalUVSkinnedVertex {
             .add(bgfx::Attrib::Position, 3, bgfx::AttribType::Float)
             .add(bgfx::Attrib::Normal, 3, bgfx::AttribType::Float)
             .add(bgfx::Attrib::TexCoord0, 2, bgfx::AttribType::Float)
-            .add(bgfx::Attrib::Indices, 4, bgfx::AttribType::Uint8)
+            .add(bgfx::Attrib::Indices, 4, bgfx::AttribType::Float)
             .add(bgfx::Attrib::Weight, 4, bgfx::AttribType::Float)
             .end();
         return l;
